refactor - graph construction extracted to GraphFactory, related types conforms
